requirements.txt
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.7
pytest>=6.2.5
